Move math functions to utils module
