+++ constants/Exercise.ts
@@ -9,19 +9,6 @@ export enum ExerciseDifficulties {
     Advanced = "Advanced"
 }
 
-export const getRndInt = (min: number, max: number) => {
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-export const getRndSign = () => {
-    return Math.random() < 0.5 ? -1 : 1;
-}
-
-// because js uses remainder not modulo :(
-export const modulo = (a: number, b: number) => {
-    return ((a % b) + b) % b;
-}
-
 export type Note = {
     midi: number,
     detune?: number,
